bug fix: convert cordinates in class star to float

+++ star.h
@@ -4,21 +4,21 @@
 
 class star {
 public:
-	star(int,int,int);
+	star(int,float,float);
 
-	friend int getDistance(star const& s1,star const& s2);
+	friend float getDistance(star const& s1,star const& s2);
 private:
 	int id;
-	int x_cord;
-	int y_cord;
+	float x_cord;
+	float y_cord;
 };
 
-star::star(int id,int x_cord,int y_cord)
+star::star(int id,float x_cord,float y_cord)
 	:id(id),x_cord(x_cord),y_cord(y_cord)
 {
 
 }
 
-int getDistance(star const& s1, star const& s2) {
+float getDistance(star const& s1, star const& s2) {
 	return sqrt(pow(s1.x_cord- s2.x_cord, 2) + pow(s1.y_cord- s1.y_cord, 2));
 }
